requirements.txt
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
kiwisolver==1.4.9
numpy==2.4.0
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.18
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
urllib3==2.5.0
